AWS_Lambda/dockerfile/requirements.txt
paho-mqtt
google-cloud-firestore
google-auth